ImmaturePointsForStereo

+++ Include/CalibData.h
@@ -86,7 +86,7 @@ class CalibData
 
     inline CalibData(int _Width, int _Height, Mat33f K, float baseline, std::shared_ptr<PhotometricUndistorter> _PhoUndL,
                      std::shared_ptr<PhotometricUndistorter> _PhoUndR, int DirPyrSize, float DirScaleFactor, int IndPyrLevels,
-                     int IndPyrScaleFactor) : Width(_Width), Height(_Height), PhotoUnDistL(_PhoUndL), PhotoUnDistR(_PhoUndL), mbf(baseline)
+                     float IndPyrScaleFactor) : Width(_Width), Height(_Height), PhotoUnDistL(_PhoUndL), PhotoUnDistR(_PhoUndL), mbf(baseline)
     {
 
         VecC initial_value = VecC::Zero();
